Migrate region tests
